Fix disallowed exports from action file as well as tests
